+++ src/aioamazondevices/const.py
@@ -54,41 +54,10 @@ CSRF_COOKIE = "csrf"
 REFRESH_ACCESS_TOKEN = "access_token"  # noqa: S105
 REFRESH_AUTH_COOKIES = "auth_cookies"
 
-NODE_DEVICES = "devices"
-NODE_DO_NOT_DISTURB = "doNotDisturbDeviceStatusList"
-NODE_PREFERENCES = "devicePreferences"
-NODE_BLUETOOTH = "bluetoothStates"
-NODE_IDENTIFIER = "identifier"
-NODE_SENSORS = "sensors"
-
-URI_QUERIES = {
-    NODE_DEVICES: "/api/devices-v2/device",
-    NODE_DO_NOT_DISTURB: "/api/dnd/device-status-list",
-    NODE_PREFERENCES: "/api/device-preferences",
-    NODE_BLUETOOTH: "/api/bluetooth",
-    # "/api/ping"
-    # "/api/np/command"
-    # "/api/np/player"
-    # "/api/device-wifi-details"
-    # "/api/activities"
-    # "/api/behaviors/v2/automations"
-    # "/api/notifications"
-}
-
+URI_DEVICES = "/api/devices-v2/device"
 URI_SIGNIN = "/ap/signin"
-URI_IDS = "/api/phoenix"
-URI_SENSORS = "/api/phoenix/state"
+URI_NEXUS_GRAPHQL = "/nexus/v1/graphql"
 
-SENSORS = [
-    "babyCryDetectionState",
-    "beepingApplianceDetectionState",
-    "coughDetectionState",
-    "dogBarkDetectionState",
-    "humanPresenceDetectionState",
-    "illuminance",
-    "temperature",
-    "waterSoundsDetectionState",
-]
 SENSOR_STATE_OFF = "NOT_DETECTED"
 
 # File extensions
@@ -100,6 +69,32 @@ BIN_EXTENSION = ".bin"
 SPEAKER_GROUP_FAMILY = "WHA"
 SPEAKER_GROUP_MODEL = "Speaker Group"
 
+SENSORS: dict[str, dict[str, str | None]] = {
+    "temperatureSensor": {
+        "name": "temperature",
+        "key": "value",
+        "subkey": "value",
+        "scale": "scale",
+    },
+    "motionSensor": {
+        "name": "motion",
+        "key": "detectionStateValue",
+        "subkey": None,
+        "scale": None,
+    },
+    "lightSensor": {
+        "name": "illuminance",
+        "key": "illuminanceValue",
+        "subkey": "value",
+        "scale": None,
+    },
+    "speaker": {
+        "name": "volume",
+        "key": "value",
+        "subkey": "volValue",
+        "scale": None,
+    },
+}
 DEVICE_TO_IGNORE: list[str] = [
     AMAZON_DEVICE_TYPE,  # Alexa App for iOS
     "A2TF17PFR55MTB",  # Alexa App for Android